requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.6
pyreadline3==3.4.1
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.1.0
